% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eqs.r
\name{cluster.df.update.var.eq.info}
\alias{cluster.df.update.var.eq.info}
\title{Update df$vars, df$vars.id, df$is.free.var, and df$is.check.eq}
\usage{
cluster.df.update.var.eq.info(df)
}
\description{
Update df$vars, df$vars.id, df$is.free.var, and df$is.check.eq
}

