% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqs.r
\name{cluster.equations}
\alias{cluster.equations}
\title{Try to solve or simplify system of equations}
\usage{
cluster.equations(eqs, var.li = NULL, endo = NULL, exo = NULL,
  verbose = TRUE, solve.symbolic = TRUE, skip.eat = FALSE,
  solve.level = 100, skip.big = FALSE, funs = NULL, cluster.df = NULL)
}
\description{
Try to solve or simplify system of equations
}

