% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subst_eqs.r
\name{suggest.eq.for.var}
\alias{suggest.eq.for.var}
\title{Suggest an equation that is easiest to solve for a variable var}
\usage{
suggest.eq.for.var(eqs, var)
}
\arguments{
\item{eqs}{The equations that are considered (list of R calls)}

\item{var}{the variable, we want to solve for (as character)}
}
\description{
Suggest an equation that is easiest to solve for a variable var
}

