% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/symsolve.r
\name{sym.solve.eq}
\alias{sym.solve.eq}
\title{Try to solve an equation symbollically}
\usage{
sym.solve.eq(eq, var)
}
\arguments{
\item{eq}{the equation as an quoted R call}

\item{var}{the variable to be solved for as a character}
}
\description{
Try to solve an equation symbollically
}

